Export app.jsx as module to support hot reloading
`docs/entry.jsx` is the new entrypoint. Requires change to archetype.
