+++ docs/victory-bar/docs.jsx
@@ -16,10 +16,11 @@ class Docs extends React.Component {
           overview={require("!!raw!./ecology.md")}
           source={docgen.parse(require("!!raw!../../src/components/victory-bar/victory-bar"))}
           scope={{_, React, ReactDOM, VictoryLabel, VictoryBar}}
-          playgroundtheme="elegant" />
+          playgroundtheme="elegant"
+        />
         <Style rules={VictoryTheme}/>
       </div>
-    )
+    );
   }
 }
